+++ planet-mitchell/src/Optic/Lens.hs
@@ -3,10 +3,6 @@ module Optic.Lens
     Lens
   , Lens'
   , lens
-    -- * HasAny
-  , HasAny(..)
-    -- * Subtype
-  , Subtype(..)
     -- * At
   , At(..)
   , sans
@@ -19,5 +15,3 @@ module Optic.Lens
 import Control.Lens.At               (At(at), Contains(contains), Index,
                                       IxValue, sans)
 import Control.Lens.Lens             (Lens, Lens', lens)
-import Data.Generics.Product.Any     (HasAny(the))
-import Data.Generics.Product.Subtype (Subtype(smash, super, upcast))